# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nyaya-ai/voice-token:
    post:
      operationId: nyayaAiVoiceToken
      tags: [nyaya-ai]
      summary: Mint an ephemeral Gemini Live session token for voice chat
      description: >
        Real-time voice conversation support. Returns a short-lived,
        single-use ephemeral token that the browser uses to open a Gemini
        Live API WebSocket directly with Google — the GEMINI_API_KEY secret
        never leaves the server. The token is constraint-locked at mint
        time: model, child-friendly voice, transcription settings and the
        ENTIRE system instruction (safety rules + the pre-approved India
        Code corpus + safe game context) are fixed server-side and cannot
        be altered by the client. No audio or transcripts are ever stored
        server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NyayaAiVoiceTokenInput"
      responses:
        "200":
          description: Ephemeral Live session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NyayaAiVoiceTokenReply"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
        "502":
          description: Upstream AI error (token could not be minted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
        "503":
          description: GEMINI_API_KEY is not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
  /nyaya-ai/voice-guard:
    post:
      operationId: nyayaAiVoiceGuard
      tags: [nyaya-ai]
      summary: Deterministic safety gate for live voice transcripts
      description: >
        Voice-mode twin of the text pipeline's deterministic gates, using
        the SAME shared safety module. The client streams Live API
        transcripts here — both the child's speech (role user) and the
        model's spoken reply (role model). When escalated=true the client
        must stop audio playback, end the Live session, show the returned
        canonical helpline text (hard-coded server-side, never
        model-generated) and pulse the shared Get Help Now button. Fully
        deterministic (no AI call), so it works even without
        GEMINI_API_KEY.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NyayaAiVoiceGuardInput"
      responses:
        "200":
          description: Guard verdict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NyayaAiVoiceGuardReply"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
...
components:
  schemas:
...
    NyayaAiGameContext:
      type: object
      description: >
        Safe, app-generated game state for personalization (PRD §9 data
        minimization: stable zone IDS — mapped to names server-side — plus
        capped counts and the fun game nickname; never any real-world PII).
      properties:
        nickname:
          type: string
          maxLength: 24
          description: Fun made-up game nickname (never a real name)
        currentZoneId:
          type: string
          maxLength: 16
          description: Zone the player is inside (e.g. zone3)
        nearbyZoneId:
          type: string
          maxLength: 16
          description: Nearest zone on the map when not inside one
        completedZoneIds:
          type: array
          maxItems: 8
          items:
            type: string
            maxLength: 16
        progressPct:
          type: number
          minimum: 0
          maximum: 100
        badgeCount:
          type: number
          minimum: 0
          maximum: 99
        currentLessonTitle:
          type: string
          maxLength: 120
        currentLevelNumber:
          type: number
          minimum: 1
          maximum: 20
        learnQuestionsAnswered:
          type: number
          minimum: 0
          maximum: 100000
          description: >
            Total recorded quiz/recap answers (learning-insights log). Lets
            Nyaya AI answer "How am I doing?" from real stats only.
        learnAccuracyPct:
          type: number
          minimum: 0
          maximum: 100
          description: >
            Overall recorded-answer accuracy — the client sends it only
            when the analyzer's minimum-evidence gate passed.
        learnTrend:
          type: string
          enum: [improving, steady, declining]
          description: Session-over-session accuracy direction (evidence-gated)
        strongZoneId:
          type: string
          maxLength: 16
          description: Zone with the strongest evidenced accuracy label
        practiceZoneId:
          type: string
          maxLength: 16
          description: Zone the deterministic analyzer suggests practicing
    NyayaAiVoiceTokenInput:
      type: object
      description: >
        Everything is optional: voice chat is reachable from the Home
        screen before onboarding (no age band yet, empty context).
      properties:
        language:
          type: string
          enum: [en, hi]
          description: >
            App UI language (defaults to en). The Live session follows the
            language the child actually SPEAKS; this is the fallback for
            mixed/unclear speech and for canonical safety text.
        ageBand:
          type: string
          enum: ["8-11", "12-15", "16-18"]
          description: Optional tone tuning only; never changes allowed facts
        gameContext:
          $ref: "#/components/schemas/NyayaAiGameContext"
    NyayaAiVoiceTokenReply:
      type: object
      properties:
        token:
          type: string
          description: Ephemeral token name the browser connects with
        model:
          type: string
          description: Live model the token is locked to
        expiresAt:
          type: string
          description: ISO time by which the session must be STARTED
      required:
        - token
        - model
        - expiresAt
    NyayaAiVoiceGuardInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 2000
          description: One finished transcript utterance to check
        role:
          type: string
          enum: [user, model]
          description: user = the child's speech, model = Nyaya AI's speech
        language:
          type: string
          enum: [en, hi]
          description: Canonical safety text language fallback (default en)
      required:
        - text
        - role
    NyayaAiVoiceGuardReply:
      type: object
      properties:
        escalated:
          type: boolean
          description: True when the deterministic safety gate fired
        reply:
          type: string
          description: >
            Canonical hard-coded escalation text (present when escalated).
      required:
        - escalated
...
    AvatarError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...