# lib/api-zod/src/generated/types/nyayaAiVoiceGuardInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NyayaAiVoiceGuardInputLanguage } from './nyayaAiVoiceGuardInputLanguage';
import type { NyayaAiVoiceGuardInputRole } from './nyayaAiVoiceGuardInputRole';

export interface NyayaAiVoiceGuardInput {
  /**
     * One finished transcript utterance to check
     * @minLength 1
     * @maxLength 2000
     */
  text: string;
  /** user = the child's speech, model = Nyaya AI's speech */
  role: NyayaAiVoiceGuardInputRole;
  /** Canonical safety text language fallback (default en) */
  language?: NyayaAiVoiceGuardInputLanguage;
}