# lib/api-zod/src/generated/types/nyayaAiVoiceGuardReply.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NyayaAiVoiceGuardReply {
  /** True when the deterministic safety gate fired */
  escalated: boolean;
  /** Canonical hard-coded escalation text (present when escalated). */
  reply?: string;
}